extract translation strings on router settings page
